% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{download_images}
\alias{download_images}
\title{download_images}
\usage{
download_images(photo_id = NULL, saveDir = "downloaded_images")
}
\arguments{
\item{photo_id}{id of pgoto to dowload, can be single id, list or column for
phot_search outputs}

\item{saveDir}{name of directory for photos to be saved in - if it doesnt
exisit it will be created}
}
\value{
jpeg image saved as the name of the photo id in given save directory
}
\description{
Downloads images based on their FLickr id. Uses the flickr.photos.getSizes
API method from the Flickr API to test whether you have permission to
download an image. See
\url{https://www.flickr.com/services/api/flickr.photos.getSizes.html} for more
information on the API method. If permission is available the image is
downloaded and saved as a .jpeg in a given save directory.
}
\details{
Note: if this is the first function of the package you use you will be prompted to enter your API
key. API keys are avialable from
\url{https://www.flickr.com/services/apps/create/apply}. The API key will
then be saved as a .Rda file and be called to when using any other function
}
\examples{
\dontrun{
download_images(photo_id = 123, saveDir = "images")

download_images(photo_id = photo_search_outputs$id, saveDir = "downloaded_photos")
}
}
