% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{downloadImages}
\alias{downloadImages}
\alias{download_images}
\title{Download images found using \code{photo_search}}
\usage{
download_images(photo_search_results = NULL, licenses = 7:10,
  saveDir = ".", max_quality = 2, verbose = TRUE)
}
\arguments{
\item{licenses}{Numeric vector, set the licenses you want to download images for. See \code{getLicenses}}

\item{saveDir}{Character, the path where images should be saved}

\item{max_quality}{Numeric 1-4 giving the maximum quality of image you want to download 1=small, 4=large}

\item{verbose}{logical, if TRUE the progeress through images is given}

\item{photoSearch_results}{A data.frame output from \code{photo_search}.}
}
\value{
A vector of all the URLs that where targetted for download
}
\description{
Using the data returned from the
}
\examples{
# run a workflow, using the logistic regression model
\dontrun{

uk_trees <- photo_search(min_taken = "2019-01-01", max_taken = "2019-01-02", text = "tree", bbox = "-13.623047,47.279229,3.251953,60.630102", has_geo = TRUE)

downloadImages(
  photo_search_results = uk_trees,
  licenses = c(6:10),
  saveDir = tempdir(),
  max_quality = 2
)
}

}
