% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting_list.R
\name{interesting_list}
\alias{interesting_list}
\title{Interesting list}
\usage{
interesting_list(date = "2019-01-01")
}
\arguments{
\item{date}{Character, required. Interestingness list for the date provided.
The date should be in the form of "YYYY-MM-DD".}
}
\description{
Returns a Flickr generated list of photographs deemed interesting. For
information on how this list is calculated see:
\link{http://www.steves-digicams.com/knowledge-center/how-tos/online-sharing-social-networking/what-is-flickr-interestingness.html#b}
}
\examples{
\dontrun{

interesting_list(date = "2019-01-01")

interesting_list(date = "2017-05-05")

interesting_list(date = "2011-11-25")

}
}
