% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_tags.R
\name{location_tags}
\alias{location_tags}
\title{location_tags}
\usage{
location_tags(woe_id = NULL)
}
\arguments{
\item{woe_id}{Numeric, a "Where on Earth" location tag.}
}
\value{
List of the top 100 tags associated with the woe_id.
}
\description{
location_tags
}
\examples{
\dontrun{
location_tags(woe_id = 35356)
}
}
