% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_search.R
\name{photo_search}
\alias{photo_search}
\title{photo_search}
\usage{
photo_search(min_taken = "2019-01-01", max_taken = "2019-01-01",
  text = NULL, bbox = NULL, has_geo = TRUE)
}
\arguments{
\item{min_taken}{String, minimum date of photograph for search provide as "YYYY-MM-DD".}

\item{max_taken}{String, maximum date of photograph for search provide as "YYYY-MM-DD".}

\item{text}{String, text to be searched.}

\item{bbox}{String, optional bounding box of search area provide as:
"minimum_longitude,minimum_latitude,maximum_longitude,maximum_latitude".}

\item{has_geo}{Logical, arguement for whether returned photos need to be georeference.}
}
\value{
Output will be a dataframe consisting of 54 variables including; latitude and longitude of photograph, photograph tags and image urls
}
\description{
photo_search
}
\examples{
\dontrun{
photo_search(min_taken = "2019-01-01",
             max_taken = "2019-01-02",
             text = "tree",
             bbox = "-13.623047,47.279229,3.251953,60.630102",
             has_geo = TRUE)

photo_search(min_taken = "2001-01",
             max_taken = "2010-01-01",
             text = "mountain",
             bbox = NULL,
             has_geo = NULL)
}
}
