% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosearcher-package.R
\docType{package}
\name{photosearcher-package}
\alias{photosearcher}
\alias{photosearcher-package}
\title{photosearcher: Photo Searcher}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Queries the Flick API (https://www.flickr.com/services/api/) to return photograph metadata as well as the ability to download the images as jpegs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/photosearcher}
  \item Report bugs at \url{https://github.com/ropensci/photosearcher/issues}
}

}
\author{
\strong{Maintainer}: Nathan Fox \email{nf2g13@soton.ac.uk}

Authors:
\itemize{
  \item Francesca Mancini
  \item Laura Graham
  \item Louis Sutter
  \item Tom August
}

}
\keyword{internal}
