% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_reddit.R
\name{reddit_search}
\alias{reddit_search}
\title{Search Reddit for posts}
\usage{
reddit_search(
  search_term = NULL,
  subreddit = NULL,
  start_date = "2020-01-01",
  end_date = "2021-01-01"
)
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
Searches for a a given term or set of terms across Reddit between chosen dates.
Can be limited to a specific subreddit.
}
