% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_tags.R
\name{related_tags}
\alias{related_tags}
\title{Get related tags}
\usage{
related_tags(tag)
}
\arguments{
\item{tag}{character. tag to search.}
}
\value{
character. Tags most associated with input tag.
}
\description{
Takes a tag and returns the top tags related to that tag.
}
\details{
Uses the flickr.tags.getRelated API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.tags.getRelated.html} for
more information on the API method.

When running the function you need an API key saved as
photosearcher_key.sysdata in your working directory. If this is the first
function you run you will be prompted to create and enter your API key. The
API key will then be saved as photosearcher_key.sysdata in your working
directory and is used for all function.
}
\examples{
\dontrun{
related_tags(tag = "car")

related_tags(tag = "monkey")

related_tags(tag = "river")
}
}
\concept{Tag information}
