% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_terms.R
\name{related_terms}
\alias{related_terms}
\title{related_terms}
\usage{
related_terms(term = NULL)
}
\arguments{
\item{term}{String, term to search.}
}
\value{
List of terms most associated with input term.
}
\description{
Takes user defined term and returns the top tags related to the term.
Uses the flickr.tags.getRelated API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.tags.getRelated.html} for
more information on the API method.
}
\details{
Note: if this is the first function of the package you use you will be
prompted to enter your API key. API keys are avialable from
\url{https://www.flickr.com/services/apps/create/apply}. The API key will
then be saved as a .Rda file and be called to when using any other function.
}
\examples{
\dontrun{
related_terms(term = "car")

related_terms(term = "monkey")

related_terms(term = "river")
}
}
