% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_key.R
\name{save_key}
\alias{save_key}
\title{save_key}
\usage{
save_key(api_key = NULL)
}
\arguments{
\item{api_key}{String, a valid API key avialable from
\url{https://www.flickr.com/services/apps/create/apply}}
}
\value{
Saves you API key as a .rda file so that it can be used for all
functions
}
\description{
Enter a valid Flickr API key.API keys are avialable from
\url{https://www.flickr.com/services/apps/create/apply}. The API key will
then be saved as a .Rda file and be called to when using any other function.
}
