% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{download_images}
\alias{download_images}
\title{download_images}
\usage{
download_images(photo_id = NULL, saveDir = "downloaded_images")
}
\arguments{
\item{photo_id}{id of pgoto to dowload, can be single id, list or column for phot_search outputs}

\item{saveDir}{name of directory for photos to be saved in - if it doesnt exisit it will be created}
}
\value{
jpeg image saved as the name of the photo id
}
\description{
download_images
}
\examples{
\dontrun{
download_images(photo_id = 123, saveDir = "images")

download_images(photo_id = photo_search_outputs$id, saveDir = "downloaded_photos")
}
}
