% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{download_images}
\alias{download_images}
\title{Download images}
\usage{
download_images(photo_id, save_dir = NULL, max_image_height = NULL,
  max_image_width = NULL, quiet = FALSE)
}
\arguments{
\item{photo_id}{numeric or character vector. id of photo to download, can be
single id, list or column for photo_search outputs}

\item{save_dir}{character. name of directory for photos to be saved in. This
will be created if it doesn't exist.}

\item{max_image_height}{numeric. maximum number of pixels for images height}

\item{max_image_width}{numeric. maximum number of pixels for images width #'}

\item{quiet}{logical. If TRUE, suppress status messages (if any), and the
progress bar.}
}
\value{
character. A vector of the images that the user does not have
permission to download. Images will be saved to \code{save_dir}.
}
\description{
Downloads images based on their Flickr ID. Uses the flickr.photos.getSizes
API method from the Flickr API to test whether you have permission to
download an image. See
\url{https://www.flickr.com/services/api/flickr.photos.getSizes.html} for
more information on the API method. If permission is available the image is
downloaded and saved as a .jpeg in a given save directory.
}
\examples{
\dontrun{
download_images(photo_id = 48704764812, save_dir = "images")

download_images(photo_id = 48704764812, maximum_image_height = 1200,
maximum_image_width = 1200, save_dir = "smaller_images") }
}
\seealso{
Other Get data for known photograph: \code{\link{get_exif}},
  \code{\link{get_photoinfo}}
}
\concept{Get data for known photograph}
