% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{download_images}
\alias{download_images}
\title{download_images}
\usage{
download_images(photo_id, saveDir = "downloaded_images", quiet = FALSE)
}
\arguments{
\item{photo_id}{numeric. id of photo to dowload, can be single id, list or column for
photo_search outputs}

\item{saveDir}{character. name of directory for photos to be saved in. This will be
created if it doesn't exist.}

\item{quiet}{logical. If TRUE, suppress status messages (if any), and the progress bar.}
}
\value{
character. A vector of the images that the user does not have permission to
download. Images will be saved to \code{saveDir}.
}
\description{
Downloads images based on their Flickr ID. Uses the flickr.photos.getSizes API method
from the Flickr API to test whether you have permission to download an image. See
\url{https://www.flickr.com/services/api/flickr.photos.getSizes.html} for more
information on the API method. If permission is available the image is downloaded and
saved as a .jpeg in a given save directory.
}
\examples{
\dontrun{
download_images(photo_id = 123, saveDir = "images")

download_images(photo_id = photo_search_outputs$id, saveDir = "downloaded_photos")
}
}
