% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_place.R
\name{findPlaces}
\alias{findPlaces}
\alias{find_place}
\title{Find correct location informatio for a given place}
\usage{
find_place(place = NULL)
}
\arguments{
\item{place}{Text string describing the place for the query}
}
\value{
A dataframe information on locations that share the name with the
search location. Full list of returned data: place_id, woeid, latitude,
longitude and place_url
}
\description{
Takes user defined location and returns location data for the search. Uses
the flickr.places.find API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.places.find.html} for more
information on the API method.
}
\details{
Note: if this is the first function of the package you use you will be
prompted to enter your API key. API keys are avialable from
\url{https://www.flickr.com/services/apps/create/apply}. The API key will
then be saved as a .Rda file and be called to when using any other function.
}
\examples{
\dontrun{
find_place(place = "New York")

find_place(place = "England")
}
}
