% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_place.R
\name{findPlaces}
\alias{findPlaces}
\alias{find_place}
\title{Find correct location informatio for a given place}
\usage{
find_place(place = NULL)
}
\arguments{
\item{place}{Text string describing the place for the query}
}
\value{
A dataframe of places including woe_id to be used in location_tags
}
\description{
Find correct location informatio for a given place
}
\examples{
\dontrun{
find_place(place = "New York")

find_place(place = "England")
}
}
