% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_place.R
\name{findPlaces}
\alias{findPlaces}
\alias{find_place}
\title{Find correct location information for a given place}
\usage{
find_place(place)
}
\arguments{
\item{place}{character. The place for the query}
}
\value{
data.frame. Information on locations that share the name with the
search location. Nine variables are returned: place_id, woeid, latitude,
longitude, place_url, place_type, place_type_id, timezone, woe_name.
}
\description{
Takes user defined location and returns location data for the search. Uses
the flickr.places.find API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.places.find.html} for more
information on the API method.
}
\examples{
\dontrun{
find_place(place = "New York")

find_place(place = "England")
}
}
