% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exif.R
\name{get_exif}
\alias{get_exif}
\title{Get Exif data}
\usage{
get_exif(photo_id = NULL)
}
\arguments{
\item{photo_id}{Id of photograph}
}
\value{
A dataframe of "exchangeable image file format" information for the
given photograph
}
\description{
Returns Exchangeable image file format data for a single photograph. For more
information on how Exif data differs from metadata see:
\link{https://www.flickr.com/services/api/flickr.photos.getExif.html}
}
\examples{
\dontrun{
get_exif(photo_id = 48704764812)
}
}
