% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exif.R
\name{get_exif}
\alias{get_exif}
\title{Get Exif data}
\usage{
get_exif(photo_id = NULL)
}
\arguments{
\item{photo_id}{Id of photograph}
}
\value{
A dataframe of "exchangeable image file format" information for the
given photograph
}
\description{
Returns Exchangeable image file format data for a single photograph. For more
information on how Exif data differs from metadata see:
\url{https://www.flickr.com/services/api/flickr.photos.getExif.html}
}
\details{
When running the function you need an API key saved as
photosearcher_key.sysdata in your working directory. If this is the first
function you run you will be prompted to create and enter your API key. The
API key will then be saved as photosearcher_key.sysdata in your working
directory and is used for all function.
}
\examples{
\dontrun{
get_exif(photo_id = 48704764812)
}
}
\seealso{
Other Get data for known photograph: \code{\link{download_images}},
  \code{\link{get_photoinfo}}
}
\concept{Get data for known photograph}
