% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting_list.R
\name{interesting_list}
\alias{interesting_list}
\title{interesting_list}
\usage{
interesting_list(date = "2019-01-01")
}
\arguments{
\item{date}{Character, required. Interestingness list for the date provided.
The date should be in the form of "YYYY-MM-DD".}
}
\description{
interesting_list
}
\examples{
\dontrun{

interesting_list(date = "2019-01-01")

interesting_list(date = "2017-05-05")

interesting_list(date = "2011-11-25")

}
}
