% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_tags.R
\name{location_tags}
\alias{location_tags}
\title{Get top tags for a location}
\usage{
location_tags(woe_id)
}
\arguments{
\item{woe_id}{numeric. a "Where on Earth" location tag.}
}
\value{
character. List of the top 100 tags associated with the woe_id.
}
\description{
Takes user defined location and returns the top tags related to the location.
Uses the flickr.places.tagsForPlace API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.places.tagsForPlace.html} for
more information on the API method.
}
\examples{
\dontrun{
location_tags(woe_id = 35356)
}
}
