% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_terms.R
\name{related_terms}
\alias{related_terms}
\title{Title}
\usage{
related_terms(term = NULL)
}
\arguments{
\item{term}{String, term to search.}
}
\value{
List of terms most associated with input term.
}
\description{
Title
}
\examples{
related_terms(term = "car")

related_terms(term = "monkey")

related_terms(term = "river")
}
