% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{user_info}
\alias{user_info}
\title{Title}
\usage{
user_info(user_id = NULL)
}
\arguments{
\item{user_id}{String, the id of the user you wihs to obtain information for.}
}
\value{
Dataframe of users publically available infromation including;
hometown and occupation
}
\description{
Title
}
\examples{
\dontrun{
user_info(user_id = "8017836@N02")

user_info(user_id = c("11763518@N00", "39745335@N06", "104939923@N02"))
}
}
