% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{user_info}
\alias{user_info}
\title{User_info}
\usage{
user_info(user_id = NULL, api_key = api_key)
}
\arguments{
\item{user_id}{String, the id of the user you wihs to obtain information for.}

\item{api_key}{String, if you have used the save_key function the api_key
argument is automatically filled. If not api_key can be used optionally to
supplying your API key if you do not wish for it to be saved in the
environment or as a .Rda}
}
\value{
Dataframe of 18 variables from the searched users publically
available infromation including; hometown and occupation. Full list of
returned variables: id, nsid,	join_date, occupation, hometown, showcase_set,
showcase_set_title, first_name, last_name, profile_description,
website	city, country, facebook, twitter, tumblr, Instagram, pinterest
}
\description{
Takes user defined FLickr user ID and returns data regarding their profile.
Uses the flickr.profile.getProfile API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.profile.getProfile.html} for
more information on the API method.
}
\details{
Note: if this is the first function of the package and you do not enter you
API key in the arguement api_key you use you will be prompted to enter your
API key or save it using the save_key function. API keys are avialable from
\url{https://www.flickr.com/services/apps/create/apply}. Using the save_key
function will save your key as a .Rda file which can then be called to when
using any other function.
}
\examples{
\dontrun{
user_info(user_id = "8017836@N02")

user_info(user_id = c("11763518@N00", "39745335@N06", "104939923@N02"))
}
}
