% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{user_info}
\alias{user_info}
\title{Get user information}
\usage{
user_info(user_id)
}
\arguments{
\item{user_id}{character. The id of the user you wish to obtain information
for.}
}
\value{
data.frame. Dataframe of 5 variables from the searched users
publicly available information: id, occupation, hometown, city, country.
}
\description{
Takes Flickr user ID and returns the profile data.
}
\details{
Uses the flickr.profile.getProfile API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.profile.getProfile.html} for
more information on the API method.
}
\examples{
\dontrun{
user_info(user_id = "155421853@N05")

user_info(user_id = photo_search_outputs$owner)
}
}
\concept{User information}
