% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{user_info}
\alias{user_info}
\title{Get user information}
\usage{
user_info(user_id)
}
\arguments{
\item{user_id}{character. The id of the user you wish to obtain information
for.}
}
\value{
data.frame. Dataframe of 5 variables from the searched users publicly
available information: id, occupation, hometown, city, country.
}
\description{
Takes Flickr user ID and returns the profile data.
}
\details{
Uses the flickr.profile.getProfile API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.profile.getProfile.html} for
more information on the API method.

See \url{https://www.pixsy.com/academy/flickr-id/} for a guide on finding
your Flickr ID.

When running the function you need an API key saved as
photosearcher_key.sysdata in your working directory. If this is the first
function you run you will be prompted to create and enter your API key. The
API key will then be saved as photosearcher_key.sysdata in your working
directory and is used for all function.
}
\examples{
\dontrun{
user_info(user_id = "155421853@N05")
}
}
\concept{User information}
