% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{user_info}
\alias{user_info}
\title{Get user information}
\usage{
user_info(user_id)
}
\arguments{
\item{user_id}{character. The id of the user you wish to obtain information for.}
}
\value{
data.frame. Dataframe of 18 variables from the searched users publically
available information: id, nsid,	join_date, occupation, hometown, showcase_set,
showcase_set_title, first_name, last_name, profile_description,
website	city, country, facebook, twitter, tumblr, Instagram, pinterest
}
\description{
Takes Flickr user ID and returns the profile data.
}
\details{
Uses the flickr.profile.getProfile API method from the Flickr API. See
\url{https://www.flickr.com/services/api/flickr.profile.getProfile.html} for
more information on the API method.
}
\examples{
\dontrun{
user_info(user_id = "8017836@N02")

user_info(user_id = c("11763518@N00", "39745335@N06", "104939923@N02"))
}
}
